Fixed bug with end-to-end tests where multiple connections accessing the same database could not operate at the same time.
The QPM was trying to release locks that it didn't hold because it was traversing all query proxies for Table Managers, instead of only traversing the query proxies with locks.

Locks are now taken out on a lockrequest object which encapsulates a database instance and a particular session so that connections to the same database can't break serializability.