Short version: Created a mechanism for re-running benchmarking output as a JUnit test.
Long version:
Added more debug that makes it easier to see what queries are being executed at runtime.

Created a script which enables console output to be parsed so that only executed queries are displayed.

Created a test program that parses the results of this script and runs queries from a test file as part of the benchmark junit tests.